Split classes out of expression.js
