+++ src/style-spec/expression/definitions/array.js
@@ -10,7 +10,9 @@ const {
     BooleanType
 } = require('../types');
 
-import type { Expression, ParsingContext, CompilationContext } from '../expression';
+import type { Expression } from '../expression';
+import type ParsingContext from '../parsing_context';
+import type CompilationContext  from '../compilation_context';
 import type { ArrayType } from '../types';
 
 const types = {
