+++ src/style-spec/expression/definitions/case.js
@@ -4,7 +4,9 @@ const assert = require('assert');
 const parseExpression = require('../parse_expression');
 const { BooleanType } = require('../types');
 
-import type { Expression, ParsingContext, CompilationContext } from '../expression';
+import type { Expression } from '../expression';
+import type ParsingContext from '../parsing_context';
+import type CompilationContext  from '../compilation_context';
 import type { Type } from '../types';
 
 type Branches = Array<[Expression, Expression]>;
