+++ src/style-spec/expression/definitions/coalesce.js
@@ -3,7 +3,9 @@
 const assert = require('assert');
 const parseExpression = require('../parse_expression');
 
-import type { Expression, ParsingContext, CompilationContext } from '../expression';
+import type { Expression } from '../expression';
+import type ParsingContext from '../parsing_context';
+import type CompilationContext  from '../compilation_context';
 import type { Type } from '../types';
 
 class Coalesce implements Expression {
