+++ src/style-spec/expression/definitions/coercion.js
@@ -8,7 +8,9 @@ const {
     NumberType,
 } = require('../types');
 
-import type { Expression, ParsingContext, CompilationContext } from '../expression';
+import type { Expression } from '../expression';
+import type ParsingContext from '../parsing_context';
+import type CompilationContext  from '../compilation_context';
 import type { Type } from '../types';
 
 const types = {
