+++ src/style-spec/expression/definitions/curve.js
@@ -7,7 +7,9 @@ const {
 } = require('../types');
 const parseExpression = require('../parse_expression');
 
-import type { Expression, ParsingContext, CompilationContext } from '../expression';
+import type { Expression } from '../expression';
+import type ParsingContext from '../parsing_context';
+import type CompilationContext  from '../compilation_context';
 import type { Type } from '../types';
 
 export type InterpolationType =
