+++ src/style-spec/expression/definitions/var.js
@@ -1,7 +1,9 @@
 // @flow
 
 import type { Type } from '../types';
-import type { Expression, ParsingContext, CompilationContext }  from '../expression';
+import type { Expression } from '../expression';
+import type ParsingContext from '../parsing_context';
+import type CompilationContext  from '../compilation_context';
 
 class Var implements Expression {
     key: string;
