+++ src/style-spec/expression/parsing_error.js
@@ -1,0 +1,13 @@
+// @flow
+
+class ParsingError extends Error {
+    key: string;
+    message: string;
+    constructor(key: string, message: string) {
+        super(message);
+        this.message = message;
+        this.key = key;
+    }
+}
+
+module.exports = ParsingError;
